Created first tdd for registration
- created new testing database
- created a class tdd to successfully post a registration
