--- api/integration_test.py
@@ -1,27 +1,0 @@
-import unittest
-import flask_testing
-import json
-from app import app, db, Course, Class, Prerequisite, Student, Completed, Registration
-
-
-class TestApp(flask_testing.TestCase):
-    app.config['SQLALCHEMY_DATABASE_URI'] = "sqlite://"
-    app.config['SQLALCHEMY_ENGINE_OPTIONS'] = {}
-    app.config['TESTING'] = True
-#Country roads take me home
-
-    def create_app(self):
-        return app
-
-    def setUp(self):
-        db.create_all()
-
-    def tearDown(self):
-        db.session.remove()
-        db.drop_all()
-
-
-#testing if it works
-
-if __name__ == '__main__':
-    unittest.main()
